Update README.md
Function prototypes

Examine the code. Notice the writeProverb function is defined befow the main function. Try running the program. You should get an 'undeclared identifier' error because main is calling writeProveb before its defintion.
Fix the error by uncommenting the prototype for writeProverb on line 4. Run the program and see that it works as expected now.